feat: fix bog of step.recovery and create new file Report.ts
