+++ packages/cli/src/utils/ConsoleReporter.ts
@@ -42,7 +42,7 @@ export class ConsoleReporter implements IReporter {
 			case TestEvent.StepSucceeded:
 				console.log(
 					chalk.green.bold('✔'),
-					chalk.grey(`Step ${stepName} passed ${timing?.toLocaleString()}ms)`),
+					chalk.grey(`Step ${stepName} passed (${timing?.toLocaleString()}ms)`),
 				)
 				break
 			case TestEvent.StepFailed:
